Move noise factors validation to higher level prepare function
